CDC output ; HighWaterMark stack monitoring
The current USB code is built off an app example. There are two CDC
coms enabled via configuration. A single task handles access to the
open CDC coms. Other tasks/components can enqueue 32 bit events for
this task to handle. The task also has read/write-completion switches
for both coms.

In this commit, tasks use a CDC print function that locks a global
WriteBuffer, writes data to it, enqueues a message to the USB task
signaling that a write is needed, then waits for the USB task to
notify it that the write was completed. The task then wipes the buffer
and releases the lock.

Otherwise, no major changes to the structure of the USB task were needed.
The second CDC com isn't being used.

--
The test_task (entered by main app task) can be used to launch a task
and then monitor its stack 'high water mark'. The high water mark is
the number of bytes free at a tasks max stack usage. These results are
printed to CDC com1 (ttyACM0) when the button is pressed. 'exec_app' is
the task under test, which is hello_world_task in this commit.

This can be used to debug potential stack problems and to minimize a
task's stack usage.

touch_ctmu is being brought in, but not really used. We may want to
reference some of this for this year's single slider.

added led(r, g, b) function. TODO: Led methods all use integers and
there is no PWM method implemented yet.

*this commit does not include some important MHC configuration changes*