Initial screensaver functionality
After some time idling on the menus, the screen backlight will be turned
off. After the backlight has been off for some amount of time, it is
turned back on and a screensaver is run for some time before being
killed (or it exits on its own). Should wake up on the D-pad right now.

I have noticed some bugs, but I think I fixed most of them. Help QCing
this would be appreciated.